Refactoring module main.py - 1: - moved the System class to a separate system.py module; - moved some of the functions from the SaleForm class to a separate sale_logic module.
