+++ main.py
@@ -1,16 +1,7 @@
-import calendar
 import datetime as dt
-import json
 import os
-import socket
 import subprocess
 import sys
-from configparser import (
-    ConfigParser,
-    NoSectionError,
-    NoOptionError,
-    MissingSectionHeaderError,
-)
 from datetime import date, timedelta
 from typing import Any, Type
 
@@ -19,19 +10,12 @@ from PySide6.QtCore import Qt, Signal
 from PySide6.QtGui import QPixmap
 from PySide6.QtWidgets import QCheckBox, QDialog, QHBoxLayout
 from PySide6.QtWidgets import QTableWidgetItem, QWidget
-from dotenv import load_dotenv
-from sqlalchemy import create_engine
-from sqlalchemy import engine, and_, select, func, update, desc, or_
-from sqlalchemy.exc import SQLAlchemyError
+from sqlalchemy import and_, select, func, update, desc, or_
 from sqlalchemy.orm import Session
 
 from db.models import Client
-from db.models import Holiday
 from db.models import Sale
 from db.models import Ticket
-from db.models import Workday
-from db.models.price import Price
-from db.models.user import User
 from design.authorization import Ui_Dialog
 from design.client import Ui_Dialog_Client
 from design.main_form import Ui_MainWindow
@@ -42,140 +26,10 @@ from files import kkt
 from files import otchet
 from files import windows
 from files.logger import logger, logger_wraps
+from sale_logic import calculate_age, calculate_ticket_type
+from system import System, config_data
 
-
-@logger_wraps(entry=True, exit=True, level="DEBUG", catch_exceptions=True)
-def load_coordinates(config_file):
-    """Функция для проверки загрузки файла с координатами, необходимыми для генерации билетов.
-
-    Параметры:
-        config_file (str): Путь к файлу конфигурации, содержащему координаты.
-
-    Возвращаемое значение:
-        dict: Словарь с координатами, содержащий следующие ключи:
-            - name (dict): Координаты имени с ключами 'x' и 'y'.
-            - surname (dict): Координаты фамилии с ключами 'x' и 'y'.
-            - age (dict): Координаты возраста с ключами 'x' и 'y'.
-            - duration (dict): Координаты продолжительности с ключами 'x' и 'y'.
-            - date (dict): Координаты даты с ключами 'x' и 'y'.
-            - guest (dict): Координаты статуса "гость" с ключами 'x' и 'y'.
-            - city (dict): Координаты города с ключами 'x' и 'y'.
-            - place (dict): Координаты места с ключами 'x' и 'y'.
-            - price (dict): Координаты цены с ключами 'x' и 'y'.
-            - ticket_type (dict): Координаты типа билета с ключами 'x' и 'y'.
-            - notes (dict): Координаты дополнительных отметок с ключами 'x' и 'y'.
-            - talents (dict): Координаты талантов с ключами 'x' и 'y'.
-            - qr_code (dict): Координаты QR-кода с ключами 'x' и 'y'.
-    """
-    logger.info("Запуск функции load_coordinates")
-    if not os.path.isfile(config_file):
-        raise FileNotFoundError(f"Файл конфигурации '{config_file}' не найден.")
-    try:
-        with open(config_file, "r", encoding="utf-8") as f:
-            data = json.load(f)
-        # Проверка, что ключ 'coordinates' существует в загруженных данных
-        if "coordinates" not in data:
-            raise KeyError("В конфигурации отсутствует ключ 'coordinates'.")
-        return data["coordinates"]
-    except json.JSONDecodeError:
-        raise ValueError(
-            "Ошибка в формате файла конфигурации. Убедитесь, что файл является корректным JSON."
-        )
-    except Exception as e:
-        raise RuntimeError(f"Ошибка при загрузке конфигурации: {e}")
-
-
-@logger_wraps(entry=True, exit=True, level="DEBUG", catch_exceptions=True)
-def read_config():
-    """Функция для загрузки параметров из файла конфигурации.
-
-    Параметры:
-        None
-
-    Возвращаемое значение:
-        dict: Словарь с параметрами конфигурации, содержащий следующие ключи:
-            - host (str): Хост базы данных.
-            - port (str): Порт базы данных.
-            - database (str): Имя базы данных.
-            - user (str): Имя пользователя базы данных.
-            - version (str): Версия программного обеспечения.
-            - log_file (str): Путь к файлу журнала.
-            - kol (str): Количество ПК.
-            - pc_1 (str): Имя первого ПК.
-            - pc_2 (str): Имя второго ПК.
-            - pinpad_path (str): Путь к терминалу.
-    """
-    logger.info("Запуск функции read_config")
-    config = ConfigParser()
-    try:
-        if not os.path.exists("config.ini"):
-            logger.error("Файл конфигурации не найден: config.ini")
-            raise FileNotFoundError("Файл конфигурации не найден.")
-        config.read("config.ini")
-        # Параметры для загрузки
-        required_keys = {
-            "host": "DATABASE",
-            "port": "DATABASE",
-            "database": "DATABASE",
-            "user": "DATABASE",
-            "version": "OTHER",
-            "log_file": "OTHER",
-            "kol": "PC",
-            "pc_1": "PC",
-            "pc_2": "PC",
-            "pinpad_path": "TERMINAL",
-        }
-        # Загружаем параметры и проверяем наличие
-        config_data = {}
-        for key, section in required_keys.items():
-            if not config.has_section(section):
-                logger.error(f"Отсутствует секция: '{section}'")
-                raise ValueError(f"Отсутствует секция: '{section}'")
-            if not config.has_option(section, key):
-                logger.error(f"Отсутствует параметр '{key}' в секции '{section}'")
-                raise ValueError(f"Отсутствует параметр '{key}' в секции '{section}'")
-            config_data[key] = config.get(section, key)
-        return config_data
-    except (NoSectionError, NoOptionError) as e:
-        raise ValueError(f"Ошибка в файле конфигурации: {e}")
-    except MissingSectionHeaderError:
-        raise ValueError("Ошибка: отсутствует заголовок секции в файле конфигурации.")
-    except Exception as e:
-        raise RuntimeError(f"Неизвестная ошибка при чтении файла конфигурации: {e}")
-
-
-# Чтение параметров из файла конфигурации
-config_data = read_config()
-host = config_data["host"]
-port = config_data["port"]
-database = config_data["database"]
-user = config_data["user"]
-software_version = config_data["version"]
-log_file = config_data["log_file"]
-kol_pc = config_data["kol"]
-pc_1 = config_data["pc_1"]
-pc_2 = config_data["pc_2"]
-
-# Загрузка переменных окружения из файла .env
-load_dotenv()
-
-# Проверка переменной окружения
-pswrd: str | None = os.getenv("DB_PASSWORD")
-if not pswrd:
-    logger.error("Переменная окружения DB_PASSWORD не установлена!")
-    raise ValueError("Переменная окружения DB_PASSWORD не установлена!")
-
-engine = create_engine(f"postgresql+psycopg2://{user}:{pswrd}@{host}:{port}/{database}")
-
-logger.add(log_file, rotation="1 MB")
-
-
-# Проверка загрузки файла с координатами
-try:
-    coordinates = load_coordinates("files/ticket_param.json")
-except (FileNotFoundError, KeyError, ValueError, RuntimeError) as e:
-    logger.error(str(e))
-    raise FileNotFoundError(f"Файл конфигурации '{coordinates}' не найден.")
+logger.add(System.log_file, rotation="1 MB")
 
 
 class AuthForm(QDialog):
@@ -188,7 +42,7 @@ class AuthForm(QDialog):
         # Добавляем логотип на форму
         pixmap = QPixmap("pylogo.png")
         self.ui.label_4.setPixmap(pixmap)
-        self.ui.label_7.setText(software_version)
+        self.ui.label_7.setText(System.software_version)
         self.ui.pushButton.clicked.connect(self.starting_the_main_form)
         self.ui.pushButton_2.clicked.connect(self.close)
 
@@ -221,13 +75,13 @@ class AuthForm(QDialog):
             # Добавляем в заголовок дополнительную информацию
             window.setWindowTitle(
                 "PyMASL ver. "
-                + software_version
+                + System.software_version
                 + ". Пользователь: "
                 + str(System.user.first_name)
                 + " "
                 + str(System.user.last_name)
                 + ". БД: "
-                + database
+                + System.database
             )
             # отображаем записи на форме продаж
             window.main_button_all_sales()
@@ -255,9 +109,9 @@ class AuthForm(QDialog):
                 logger.info("Сегодня день многодетных")
                 System.sunday = 1
             # считываем количество РМ и имена
-            System.kol_pc = kol_pc
-            System.pc_1 = pc_1
-            System.pc_2 = pc_2
+            System.kol_pc = System.kol_pc
+            System.pc_1 = System.pc_1
+            System.pc_2 = System.pc_2
             window.exec_()
         else:
             logger.warning(f"Неудачная авторизация пользователя: {login}")
@@ -331,7 +185,7 @@ class ClientForm(QDialog):
         if len(System.last_name) >= 2 and len(first_name) >= 2:
             if System.client_update != 1:
                 logger.info("Добавляем нового клиента")
-                with Session(engine) as session:
+                with Session(System.engine) as session:
                     # # получаем максимальный id в таблице клиентов
                     query = func.max(Client.id)
                     client_index: int = session.execute(query).scalars().first()
@@ -354,7 +208,7 @@ class ClientForm(QDialog):
             elif System.client_update == 1:
                 # обновляем информацию о клиенте
                 logger.info("Обновляем информацию о клиенте")
-                with Session(engine) as session:
+                with Session(System.engine) as session:
                     session.execute(
                         update(Client)
                         .where(Client.id == System.client_id)
@@ -433,7 +287,7 @@ class SaleForm(QDialog):
         row_number: int = self.ui.tableWidget.currentRow()
         # Получаем содержимое ячейки
         client_id: str = self.ui.tableWidget.item(row_number, 5).text()
-        with Session(engine) as session:
+        with Session(System.engine) as session:
             search_client: Type[Client] = (
                 session.query(Client).filter_by(id=client_id).first()
             )
@@ -479,7 +333,7 @@ class SaleForm(QDialog):
         """
         logger.info("Запуск функции adding_new_client_to_sale")
         # Запрашиваем данные нового клиента
-        with Session(engine) as session:
+        with Session(System.engine) as session:
             search_client = (
                 session.query(Client)
                 .filter_by(id=System.id_new_client_in_sale)
@@ -491,9 +345,9 @@ class SaleForm(QDialog):
             return
         logger.debug(f"Найденный клиент: {search_client}")
         # Вычисляем возраст клиента
-        age: int = System.calculate_age(search_client.birth_date)
+        age: int = calculate_age(search_client.birth_date) #System.calculate_age(search_client.birth_date)
         # Определяем тип билета и цену
-        type_ticket: str = System.calculate_ticket_type(age)
+        type_ticket: str = calculate_ticket_type(age)
         # Определяем тип билета и цену
         time: int = int(self.ui.comboBox.currentText())
         if type_ticket == "бесплатный":
@@ -814,13 +668,13 @@ class SaleForm(QDialog):
         if index == 2:
             # Поиск по номеру телефона
             user_filter: str = "%" + self.ui.lineEdit.text() + "%"
-            with Session(engine) as session:
+            with Session(System.engine) as session:
                 search = (
                     session.query(Client).filter(Client.phone.ilike(user_filter)).all()
                 )
             logger.debug(search)
             for client in search:
-                age = System.calculate_age(client.birth_date)
+                age = calculate_age(client.birth_date)
                 self.filling_client_table_widget(
                     client.last_name,
                     client.first_name,
@@ -833,12 +687,12 @@ class SaleForm(QDialog):
             # Поиск по фамилии
             query: str = "%" + self.ui.lineEdit.text() + "%"
             self.ui.tableWidget.setRowCount(0)
-            with Session(engine) as session:
+            with Session(System.engine) as session:
                 search: list[Type[Client]] = (
                     session.query(Client).filter(Client.last_name.ilike(query)).all()
                 )
             for client in search:
-                age = System.calculate_age(client.birth_date)
+                age = calculate_age(client.birth_date)
                 self.filling_client_table_widget(
                     client.last_name,
                     client.first_name,
@@ -854,7 +708,7 @@ class SaleForm(QDialog):
             # Разбиваем поисковую фразу на две
             lst: Any = search.split()
             if len(lst) == 2:
-                with Session(engine) as session:
+                with Session(System.engine) as session:
                     search: list[Type[Client]] = (
                         session.query(Client)
                         .filter(
@@ -872,7 +726,7 @@ class SaleForm(QDialog):
                     "",
                 )
             for client in search:
-                age = System.calculate_age(client.birth_date)
+                age = calculate_age(client.birth_date)
                 self.filling_client_table_widget(
                     client.last_name,
                     client.first_name,
@@ -884,14 +738,14 @@ class SaleForm(QDialog):
         elif index == 3:
             # Поиск инвалидов
             self.ui.tableWidget.setRowCount(0)
-            with Session(engine) as session:
+            with Session(System.engine) as session:
                 search: list[Type[Client]] = (
                     session.query(Client)
                     .filter(Client.privilege.ilike("%" + "и"))
                     .all()
                 )
             for client in search:
-                age = System.calculate_age(client.birth_date)
+                age = calculate_age(client.birth_date)
                 self.filling_client_table_widget(
                     client.last_name,
                     client.first_name,
@@ -903,14 +757,14 @@ class SaleForm(QDialog):
         elif index == 4:
             # Поиск многодетных
             self.ui.tableWidget.setRowCount(0)
-            with Session(engine) as session:
+            with Session(System.engine) as session:
                 search: list[Type[Client]] = (
                     session.query(Client)
                     .filter(Client.privilege.ilike("%" + "м"))
                     .all()
                 )
             for client in search:
-                age = System.calculate_age(client.birth_date)
+                age = calculate_age(client.birth_date)
                 self.filling_client_table_widget(
                     client.last_name,
                     client.first_name,
@@ -959,14 +813,14 @@ class SaleForm(QDialog):
         logger.warning(f"Статус продажи - новая: {System.sale_status}")
         row_number: int = self.ui.tableWidget.currentRow()
         res: str = self.ui.tableWidget.item(row_number, 5).text()
-        with Session(engine) as session:
+        with Session(System.engine) as session:
             search_client: Type[Client] = (
                 session.query(Client).filter_by(id=res).first()
             )
         # Вычисляем возраст клиента
         age: int = int(self.ui.tableWidget.item(row_number, 2).text())
         # Определяем тип билета и цену
-        type_ticket: str = System.calculate_ticket_type(age)
+        type_ticket: str = calculate_ticket_type(age)
         # Создаем виджет checkbox для скидки
         widget = QWidget()
         checkbox = QCheckBox()
@@ -1006,7 +860,7 @@ class SaleForm(QDialog):
             self.ui.tableWidget_3.removeRow(0)
         # Ищем продажи, в которых клиент был ранее
         client_list: set = set()
-        with Session(engine) as session:
+        with Session(System.engine) as session:
             sales = (
                 session.query(Client.id, Ticket.id_sale, Ticket.id_sale)
                 .filter(
@@ -1032,7 +886,7 @@ class SaleForm(QDialog):
         for client in client_list:
             search_cl_in_sale = session.query(Client).filter_by(id=client).one()
             row: int = self.ui.tableWidget_3.rowCount()
-            age: int = System.calculate_age(search_cl_in_sale.birth_date)
+            age: int = calculate_age(search_cl_in_sale.birth_date)
             self.filling_client_table_widget_3(
                 row,
                 search_cl_in_sale.last_name,
@@ -1064,10 +918,10 @@ class SaleForm(QDialog):
         res: str = self.ui.tableWidget_3.item(row_number, 4).text()
         age: int = int(self.ui.tableWidget_3.item(row_number, 2).text())
         # Находим выделенного в таблице клиента
-        with Session(engine) as session:
+        with Session(System.engine) as session:
             search_client = session.query(Client).filter_by(id=res).first()
         # Определяем тип билета и цену
-        type_ticket: str = System.calculate_ticket_type(age)
+        type_ticket: str = calculate_ticket_type(age)
         # Создаём виджет checkbox для скидки
         widget = QWidget()
         checkbox = QCheckBox()
@@ -1347,6 +1201,8 @@ class SaleForm(QDialog):
                 price: int = System.price["ticket_free"]
                 System.count_number_of_visitors["kol_adult_invalid"] += 1
                 self.update_sale_dict_adult_invalid(row)
+                # Меняем категорию билета на 'с' - сопровождающий
+                self.ui.tableWidget_2.setItem(row, 4, QTableWidgetItem("с"))
             else:
                 price: int = System.price["ticket_adult_3"]
         self.bind_adult_to_sale(row)
@@ -1448,8 +1304,8 @@ class SaleForm(QDialog):
             "kol_adult_invalid"
         ]
         System.sale_dict["detail"][1] = System.price["ticket_free"]
-        # Меняем категорию билета на 'с' - сопровождающий
-        self.ui.tableWidget_2.setItem(row, 4, QTableWidgetItem("с"))
+        # # Меняем категорию билета на 'с' - сопровождающий
+        # self.ui.tableWidget_2.setItem(row, 4, QTableWidgetItem("с"))
 
     def update_sale_dict_child_many_child(self) -> None:
         """
@@ -1793,11 +1649,11 @@ class SaleForm(QDialog):
             logger.debug(f"Получена продажа: {add_sale}")
             # Сохраняем продажу
             logger.debug("Сохраняем продажу в БД")
-            with Session(engine) as session:
+            with Session(System.engine) as session:
                 session.add(add_sale)
                 session.commit()
             # Получаем номер сохраненной продажи
-            with Session(engine) as session:
+            with Session(System.engine) as session:
                 query = func.max(Sale.id)
                 System.sale_id = session.execute(query).scalars().first()
             # Сохраняем билеты
@@ -1821,7 +1677,7 @@ class SaleForm(QDialog):
                     ticket_type=type_ticket,
                     client_age=System.sale_tickets[i][6],
                 )
-                with Session(engine) as session:
+                with Session(System.engine) as session:
                     session.add(add_ticket)
                     session.commit()
             self.close()
@@ -1872,7 +1728,7 @@ class SaleForm(QDialog):
                     else:
                         check = kkt.read_pinpad_file(config_data, remove_newline=False)
                     logger.debug(f"Чек прихода: {check}")
-                    with Session(engine) as session:
+                    with Session(System.engine) as session:
                         session.execute(
                             update(Sale)
                             .where(Sale.id == System.sale_id)
@@ -1903,7 +1759,7 @@ class SaleForm(QDialog):
                     )
                 # Обновляем информацию о продаже в БД
                 logger.debug(f"Обновляем информацию в БД о продаже: {System.sale_id}")
-                with Session(engine) as session:
+                with Session(System.engine) as session:
                     session.execute(
                         update(Sale)
                         .where(Sale.id == System.sale_id)
@@ -1947,7 +1803,7 @@ class SaleForm(QDialog):
         balance_error: int = 0
         tickets = self.generating_items_for_the_return_check()
         logger.info("Запрашиваем информацию о продаже в БД")
-        with Session(engine) as session:
+        with Session(System.engine) as session:
             query = select(Sale).filter(Sale.id == System.sale_id)
             sale = session.execute(query).scalars().one()
         logger.debug(f"Итоговая сумма: {sale.price}, тип оплаты: {sale.payment_type}")
@@ -2005,7 +1861,7 @@ class SaleForm(QDialog):
                                 config_data, remove_newline=False
                             )
                             logger.debug(f"Чек возврата: {check}")
-                            with Session(engine) as session:
+                            with Session(System.engine) as session:
                                 session.execute(
                                     update(Sale)
                                     .where(Sale.id == System.sale_id)
@@ -2055,7 +1911,7 @@ class SaleForm(QDialog):
                     else:
                         status_info: int = 2
                     logger.info("Обновляем информацию о возврате в БД")
-                    with Session(engine) as session:
+                    with Session(System.engine) as session:
                         session.execute(
                             update(Sale)
                             .where(Sale.id == System.sale_id)
@@ -2083,7 +1939,7 @@ class SaleForm(QDialog):
                     logger.info("Операция повторного возврата прошла успешно")
                     check = kkt.read_pinpad_file(config_data, remove_newline=False)
                     logger.debug(f"Чек возврата: {check}")
-                    with Session(engine) as session:
+                    with Session(System.engine) as session:
                         session.execute(
                             update(Sale)
                             .where(Sale.id == System.sale_id)
@@ -2125,7 +1981,7 @@ class SaleForm(QDialog):
         # Счетчик для отслеживания корректности проведения возврата за наличный способ расчета
         tickets = self.generating_items_for_the_return_check()
         logger.info("Запрашиваем информацию о продаже в БД")
-        with Session(engine) as session:
+        with Session(System.engine) as session:
             query = select(Sale).filter(Sale.id == System.sale_id)
             sale = session.execute(query).scalars().one()
         logger.debug(f"Итоговая сумма: {sale.price}, тип оплаты: {sale.payment_type}")
@@ -2156,7 +2012,7 @@ class SaleForm(QDialog):
                                 config_data, remove_newline=False
                             )
                             logger.debug(f"Чек возврата: {check}")
-                            with Session(engine) as session:
+                            with Session(System.engine) as session:
                                 session.execute(
                                     update(Sale)
                                     .where(Sale.id == System.sale_id)
@@ -2193,7 +2049,7 @@ class SaleForm(QDialog):
                     # Изменяем статус для отмененного платежа
                     status_info: int = 8
                     logger.info("Обновляем информацию об отмене в БД")
-                    with Session(engine) as session:
+                    with Session(System.engine) as session:
                         session.execute(
                             update(Sale)
                             .where(Sale.id == System.sale_id)
@@ -2546,7 +2402,7 @@ class SlipForm(QDialog):
             None: Функция не возвращает значений, вставляет фамилию в поле формы.
         """
         logger.info("Запуск функции get_slip")
-        with Session(engine) as session:
+        with Session(System.engine) as session:
             query = select(Sale.bank_pay).where(Sale.id == System.sale_id)
             load_slip = session.execute(query).scalars().one()
         words = load_slip.split()
@@ -2708,7 +2564,7 @@ class MainWindow(QtWidgets.QMainWindow, Ui_MainWindow):
         if index == 2:
             # Поиск по номеру телефона
             self.ui.tableWidget.setRowCount(0)
-            with Session(engine) as session:
+            with Session(System.engine) as session:
                 search: list[Type[Client]] = (
                     session.query(Client).filter(Client.phone.ilike(user_filter)).all()
                 )
@@ -2727,7 +2583,7 @@ class MainWindow(QtWidgets.QMainWindow, Ui_MainWindow):
         elif index == 1:
             # Поиск по фамилии
             self.ui.tableWidget.setRowCount(0)
-            with Session(engine) as session:
+            with Session(System.engine) as session:
                 search: list[Type[Client]] = (
                     session.query(Client)
                     .filter(Client.last_name.ilike(user_filter))
@@ -2752,7 +2608,7 @@ class MainWindow(QtWidgets.QMainWindow, Ui_MainWindow):
             # Разбиваем поисковую фразу на две
             lst: Any = user_filter.split()
             if len(lst) == 2:
-                with Session(engine) as session:
+                with Session(System.engine) as session:
                     search: list[Type[Client]] = (
                         session.query(Client)
                         .filter(
@@ -2784,7 +2640,7 @@ class MainWindow(QtWidgets.QMainWindow, Ui_MainWindow):
         elif index == 3:
             # Поиск инвалидов
             self.ui.tableWidget.setRowCount(0)
-            with Session(engine) as session:
+            with Session(System.engine) as session:
                 search: list[Type[Client]] = (
                     session.query(Client)
                     .filter(Client.privilege.ilike("%" + "и"))
@@ -2805,7 +2661,7 @@ class MainWindow(QtWidgets.QMainWindow, Ui_MainWindow):
         elif index == 4:
             # Поиск многодетных
             self.ui.tableWidget.setRowCount(0)
-            with Session(engine) as session:
+            with Session(System.engine) as session:
                 search: list[Type[Client]] = (
                     session.query(Client)
                     .filter(Client.privilege.ilike("%" + "м"))
@@ -2847,7 +2703,7 @@ class MainWindow(QtWidgets.QMainWindow, Ui_MainWindow):
         if client_id is None:
             logger.error("Ячейка пуста или не существует.")
             return
-        with Session(engine) as session:
+        with Session(System.engine) as session:
             search_client: Type[Client] = (
                 session.query(Client).filter_by(id=client_id).first()
             )
@@ -2916,7 +2772,7 @@ class MainWindow(QtWidgets.QMainWindow, Ui_MainWindow):
             row_number: int = idx.row()
             # Получаем содержимое ячейки
             sale_number: str = self.ui.tableWidget_2.item(row_number, 0).text()
-            with Session(engine) as session:
+            with Session(System.engine) as session:
                 query = (
                     select(
                         Client.last_name,
@@ -3176,7 +3032,7 @@ class MainWindow(QtWidgets.QMainWindow, Ui_MainWindow):
             dt1: str = self.ui.dateEdit_3.date().toString("yyyy-MM-dd") + start_time
             dt2: str = self.ui.dateEdit_3.date().toString("yyyy-MM-dd") + end_time
             if self.ui.checkBox.isChecked():
-                with Session(engine) as session:
+                with Session(System.engine) as session:
                     sales = (
                         session.query(
                             Sale.id,
@@ -3206,7 +3062,7 @@ class MainWindow(QtWidgets.QMainWindow, Ui_MainWindow):
                         .order_by(desc(Sale.id))
                     )
             else:
-                with Session(engine) as session:
+                with Session(System.engine) as session:
                     sales = (
                         session.query(
                             Sale.id,
@@ -3240,7 +3096,7 @@ class MainWindow(QtWidgets.QMainWindow, Ui_MainWindow):
             elif self.ui.radioButton_3.isChecked():
                 filter_day = dt.datetime.today() - timedelta(days=7)
             if self.ui.checkBox.isChecked():
-                with Session(engine) as session:
+                with Session(System.engine) as session:
                     sales = (
                         session.query(
                             Sale.id,
@@ -3269,7 +3125,7 @@ class MainWindow(QtWidgets.QMainWindow, Ui_MainWindow):
                         .order_by(desc(Sale.id))
                     )
             else:
-                with Session(engine) as session:
+                with Session(System.engine) as session:
                     sales = (
                         session.query(
                             Sale.id,
@@ -3430,7 +3286,7 @@ class MainWindow(QtWidgets.QMainWindow, Ui_MainWindow):
         end_time: str = " 23:59:59"
         dt1: str = self.ui.dateEdit_2.date().toString("yyyy-MM-dd") + start_time
         dt2: str = self.ui.dateEdit.date().toString("yyyy-MM-dd") + end_time
-        with Session(engine) as session:
+        with Session(System.engine) as session:
             query = select(
                 Sale.pc_name, Sale.payment_type, Sale.price, Sale.status
             ).where(and_(Sale.status == "1", Sale.datetime.between(dt1, dt2)))
@@ -3476,7 +3332,7 @@ class MainWindow(QtWidgets.QMainWindow, Ui_MainWindow):
         cash: int = int(pc_1["cash"]) + int(pc_2["cash"])
         summa: int = card + cash
         # Считаем возвраты
-        with Session(engine) as session:
+        with Session(System.engine) as session:
             query = select(
                 Sale.pc_name, Sale.payment_type, Sale.price, Sale.status
             ).where(
@@ -3574,7 +3430,7 @@ class MainWindow(QtWidgets.QMainWindow, Ui_MainWindow):
             2, 6, QTableWidgetItem(f"{pc_1_return + pc_2_return}")
         )
         # Считаем оплаченные билеты
-        with Session(engine) as session:
+        with Session(System.engine) as session:
             query = select(
                 Ticket.ticket_type,
                 Ticket.arrival_time,
@@ -3958,276 +3814,11 @@ class Payment:
     Offline: int = 100
 
 
-class System:
-    """Класс для хранения системной информации и функций"""
-
-    # Данные успешно авторизованного пользователя
-    user: User | None = None
-    # Флаг для обновления клиента
-    client_id: int | None = None
-    client_update: int | None = None
-    # Сохраняем фамилию нового клиента
-    last_name: str = ""
-    # Статус продажи: 0 - создана, 1 - оплачена, 2 - возвращена,
-    # 3 - требуется повторный возврат по банковскому терминалу,
-    # 4 - повторный возврат по банковскому терминалу,
-    # 5 - требуется частичный возврат, 6 - частичный возврат
-    # 7 - возврат по банковским реквизитам
-    # 8 - отменена
-    sale_status: int | None = None
-    sale_id: int | None = None
-    sale_discount: int | None = None
-    sale_tickets = ()
-    sale_tuple: tuple = ()
-    sale_special: int | None = None
-    # Номер строки с активным CheckBox для исключения взрослого из продажи
-    sale_checkbox_row: int | None = None
-    # Состояние CheckBox для искл. взрослого из продажи: 0 - есть, 1 - нет
-    exclude_from_sale: int = 0
-    # Какой сегодня день: 0 - будний, 1 - выходной
-    what_a_day: int | None = None
-    # Первое воскресенье месяца (день многодлетных): 0 - нет, 1 - да
-    sunday: int | None = None
-    today: date = date.today()
-    # Номер дня недели
-    num_of_week: int = 0
-    pc_name: str = socket.gethostname()
-    # Прайс-лист основных услуг
-    price: dict = {
-        "ticket_child_1": 0,
-        "ticket_child_2": 0,
-        "ticket_child_3": 0,
-        "ticket_child_week_1": 0,
-        "ticket_child_week_2": 0,
-        "ticket_child_week_3": 0,
-        "ticket_adult_1": 0,
-        "ticket_adult_2": 0,
-        "ticket_adult_3": 0,
-        "ticket_free": 0,
-    }
-    price_service: dict = {}
-    # Количество начисляемых талантов
-    talent: dict = {"1_hour": 25, "2_hour": 35, "3_hour": 50}
-    # Возраст посетителей
-    age: dict = {"min": 5, "max": 15}
-    # Информация о РМ
-    kol_pc: int = 0
-    pc_1: str = ""
-    pc_2: str = ""
-    sale_dict: dict = {
-        "kol_adult": 0,
-        "price_adult": 0,
-        "kol_child": 0,
-        "price_child": 0,
-        "detail": [0, 0, 0, 0, 0, 0, 0, 0],
-    }
-    # Содержание detail: [kol_adult, price_adult, kol_child, price_child, discount, id_adult, time, sum]
-
-    # Храним id нового клиента
-    id_new_client_in_sale: int = 0
-    # Флаг печати кассовых и банковских чеков
-    print_check: int = 1
-    # Сохраняем сумму для внесения или выплаты из кассы
-    amount_to_pay_or_deposit: int = 0
-    # Сохраняем баланс наличных денег в кассе
-    amount_of_money_at_the_cash_desk: int = None
-    # Словарь count_number_of_visitors -> Используем для подсчета количества посетителей: 'kol_adult', 'kol_child'
-    # Используем для подсчета количества посетителей со скидкой: 'kol_sale_adult', 'kol_sale_child'
-    # Используем для подсчета количества посетителей с категорией: 'kol_adult_many_child', 'kol_child_many_child',
-    # 'kol_adult_invalid', 'kol_child_invalid' = 0
-    # Запоминаем id для "привязки" продажи ко взрослому: 'id_adult'
-    # Флаг "многодетные" (1 - два часа бесплатно, 2 - скидка 50%): 'many_child'
-    # Считаем количество золотых талантов: 'talent'
-    count_number_of_visitors: dict = {
-        "kol_adult": 0,
-        "kol_child": 0,
-        "kol_sale_adult": 0,
-        "kol_sale_child": 0,
-        "kol_adult_many_child": 0,
-        "kol_child_many_child": 0,
-        "kol_adult_invalid": 0,
-        "kol_child_invalid": 0,
-        "id_adult": 0,
-        "many_child": 0,
-        "invalid": 0,
-        "talent": 0,
-    }
-
-    def user_authorization(self, login, password) -> int:
-        """
-        Функция проверяет есть ли пользователь в БД с данными, указанными на форме авторизации.
-
-        Параметры:
-            login (str): Логин пользователя.
-            password (str): Пароль пользователя.
-
-        Возвращаемое значение:
-            int: 1 - успешная авторизация, 0 - неуспешная.
-        """
-        logger.info("Запуск функции user_authorization")
-        try:
-            with Session(engine) as session:
-                query = select(User).where(
-                    User.login == login, User.password == password
-                )
-                kassir = session.execute(query).scalars().first()
-            if kassir:
-                System.user = kassir
-                logger.info(f"Успешная авторизация: {kassir.last_name}")
-                return 1
-            else:
-                logger.warning(
-                    f"Неудачная попытка авторизации для пользователя {login}"
-                )
-                return 0
-        except SQLAlchemyError as e:
-            logger.error(f"Ошибка базы данных при авторизации пользователя: {e}")
-            return 0
-        except Exception as e:
-            logger.error(f"Неизвестная ошибка при авторизации: {e}")
-            return 0
-
-    @logger_wraps(entry=True, exit=True, level="DEBUG", catch_exceptions=True)
-    def get_price(self) -> None:
-        """
-        Функция загружает прайс-лист основных услуг из БД.
-
-        Параметры:
-        self: object
-            Ссылка на экземпляр текущего объекта класса, необходимая для доступа к элементам GUI.
-
-        Возвращаемое значение:
-            None: Функция не возвращает значений, вставляет фамилию в поле формы.
-        """
-        logger.info("Запуск функции get_price")
-
-        # Словарь со значениями по умолчанию
-        default_prices = {
-            "ticket_child_1": 250,
-            "ticket_child_2": 500,
-            "ticket_child_3": 750,
-            "ticket_child_week_1": 300,
-            "ticket_child_week_2": 600,
-            "ticket_child_week_3": 900,
-            "ticket_adult_1": 150,
-            "ticket_adult_2": 200,
-            "ticket_adult_3": 250,
-        }
-
-        with Session(engine) as session:
-            result = session.query(Price).order_by(Price.id).all()
-
-        if result:
-            logger.debug(f"result: {result}")
-            # Словарь для хранения цен из БД
-            db_prices = {
-                "ticket_child_1": result[0],
-                "ticket_child_2": result[1],
-                "ticket_child_3": result[2],
-                "ticket_child_week_1": result[3],
-                "ticket_child_week_2": result[4],
-                "ticket_child_week_3": result[5],
-                "ticket_adult_1": result[6],
-                "ticket_adult_2": result[7],
-                "ticket_adult_3": result[8],
-            }
-
-            # Устанавливаем цены, используя значения из БД или по умолчанию
-            for key, value in db_prices.items():
-                System.price[key] = (
-                    int(str(value)) if int(str(value)) != 0 else default_prices[key]
-                )
-        else:
-            # Если результат запроса пуст, устанавливаем значения по умолчанию
-            logger.info(f"Устанавливаем значения прайс-листа по умолчанию.")
-            System.price.update(default_prices)
-        logger.debug(f"System.price: {System.price}")
-
-    def check_day(self) -> int:
-        """
-        Функция проверяет статус текущего дня.
-
-        Параметры:
-        self: object
-            Ссылка на экземпляр текущего объекта класса, необходимая для доступа к элементам GUI.
-
-        Возвращаемое значение:
-            int: 1 - выходной день, 0 - будний день.
-        """
-        logger.info("Запуск функции check_day")
-        day: list[str] = dt.datetime.now().strftime("%Y-%m-%d")
-        # Проверяем есть ли текущая дата в списке дополнительных рабочих дней
-        with Session(engine) as session:
-            query = select(Workday).where(Workday.date == day)
-            check_day: Workday | None = session.execute(query).scalars().first()
-        if check_day:
-            logger.info("Сегодня дополнительный рабочий день")
-            status_day: int = 0
-        else:
-            # Преобразуем текущую дату в список
-            day: list[str] = day.split("-")
-            # Вычисляем день недели
-            number_day: int = calendar.weekday(int(day[0]), int(day[1]), int(day[2]))
-            # Проверяем день недели равен 5 или 6
-            if number_day >= 5:
-                status_day: int = 1
-                System.what_a_day = 1  # TODO: нужно ли
-                logger.info("Сегодня выходной день")
-            else:
-                day: str = "-".join(day)
-                # Проверяем есть ли текущая дата в списке дополнительных праздничных дней
-                with Session(engine) as session:
-                    query = select(Holiday).where(Holiday.date == day)
-                    check_day: Holiday | None = session.execute(query).scalars().first()
-                if check_day:
-                    status_day: int = 1
-                    System.what_a_day = 1  # TODO: нужно ли
-                    logger.info("Сегодня дополнительный выходной")
-                else:
-                    status_day: int = 0
-        return status_day
-
-    @staticmethod
-    def calculate_age(born: date) -> int:
-        """
-        Функция вычисляет возраст посетителя.
-
-        Параметры:
-            born (date): Дата рождения.
-
-        Возвращаемое значение:
-            int: Возраст.
-        """
-        today: date = date.today()
-        return (
-            today.year - born.year - ((today.month, today.day) < (born.month, born.day))
-        )
-
-    @staticmethod
-    def calculate_ticket_type(age: int) -> str:
-        """
-        Функция определяет тип входного билета.
-
-        Параметры:
-            age (int): Возраст посетителя.
-
-        Возвращаемое значение:
-            str: Тип билета (бесплатный, детский, взрослый).
-        """
-        result: str = ""
-        if age < System.age["min"]:
-            result = "бесплатный"
-        elif System.age["min"] <= age < System.age["max"]:
-            result = "детский"
-        elif age >= System.age["max"]:
-            result = "взрослый"
-        return result
-
 
 if __name__ == "__main__":
     app = QtWidgets.QApplication(sys.argv)
 
     auth = AuthForm()
     auth.show()
-    auth.ui.label_9.setText(database)
+    auth.ui.label_9.setText(System.database)
     sys.exit(app.exec())
